Switched to 120 degree longitude separations. Some tests now failing. Expected values correct; mid way through debugging actuals.

+++ improver/utilities/spatial.py
@@ -300,8 +300,8 @@ class LatLonCubeDistanceCalculator(BaseDistanceCalculator):
         lon_diffs = np.diff(self.longs)
 
         x_distances = (
-            self.sphere_radius * np.cos(np.deg2rad(lats_as_col)) * np.deg2rad(lon_diffs)
-        )
+            self.sphere_radius * np.cos(np.deg2rad(lats_as_col.astype(np.float64))) * np.deg2rad(lon_diffs.astype(np.float64))
+        )  # Using 64 bit floats for this calculation improves precision by 0.1% TODO: check this.
 
         dims = [(self.x_diff.coord("latitude"), 0), (self.x_diff.coord("longitude"), 1)]
         return self.build_distances_cube(x_distances, dims, "x")
